+++ src/binom/sources/structure/path.cpp
@@ -227,4 +227,4 @@ bool Path::iterator::operator<(Path::iterator other) {return ptr < other.ptr;}
 
 bool Path::iterator::operator<=(Path::iterator other) {return ptr <= other.ptr;}
 
-Path binom::operator""_pth(const char* str, unsigned long long) {return Path::fromString(str);}
+Path binom::operator""_pth(const char* str, size_t) {return Path::fromString(str);}
